wallet: split transaction builder
