+++ xelis_wallet/src/lib.rs
@@ -6,6 +6,7 @@ pub mod daemon_api;
 pub mod network_handler;
 pub mod entry;
 pub mod mnemonics;
+pub mod transaction_builder;
 
 #[cfg(feature = "api_server")]
 pub mod api;
