+++ xelis_wallet/src/wallet.rs
@@ -1,5 +1,4 @@
 use std::{
-    collections::{HashMap, HashSet},
     fs::{create_dir_all, File},
     io::{Read, Write},
     path::Path,
@@ -16,7 +15,6 @@ use tokio::sync::{
     RwLock
 };
 use xelis_common::{
-    account::CiphertextCache,
     api::{
         wallet::{
             BalanceChanged,
@@ -38,9 +36,7 @@ use xelis_common::{
     network::Network,
     transaction::{
         builder::{
-            AccountState,
             FeeBuilder,
-            FeeHelper,
             TransactionBuilder,
             TransactionTypeBuilder
         },
@@ -67,10 +63,10 @@ use crate::{
         SharedNetworkHandler
     },
     storage::{
-        Balance,
         EncryptedStorage,
         Storage
-    }
+    },
+    transaction_builder::{EstimateFeesState, TransactionBuilderState}
 };
 use chacha20poly1305::{
     aead::OsRng,
@@ -703,14 +699,7 @@ impl Wallet {
         let transaction = builder.build(&mut state, &self.keypair)
             .map_err(|e| WalletError::Any(e.into()))?;
 
-        // Increment the nonce
-        storage.set_nonce(state.get_nonce()?)?;
-
-        // Update our storage in case of next transaction
-        for (asset, balance) in state.balances {
-            // We store every balances used in the transaction
-            storage.set_balance_for(&asset, balance).await?;
-        }
+        state.apply_changes(storage).await?;
 
         Ok(transaction)
     }
@@ -968,120 +957,5 @@ impl XSWDNodeMethodHandler for Arc<Wallet> {
         } else {
             Err(RpcResponseError::new(id, InternalRpcError::CustomStr("Wallet is not in online mode")))
         }
-    }
-}
-
-// State used to estimate fees for a transaction
-// Because fees can be higher if a destination account is not registered
-// We need to give this information during the estimation of fees
-pub struct EstimateFeesState {
-    // this is containing the registered keys that we are aware of
-    registered_keys: HashSet<PublicKey>
-}
-
-impl EstimateFeesState {
-    pub fn new() -> Self {
-        Self {
-            registered_keys: HashSet::new()
-        }
-    }
-
-    pub fn set_registered_keys(&mut self, registered_keys: HashSet<PublicKey>) {
-        self.registered_keys = registered_keys;
-    }
-
-    pub fn add_registered_key(&mut self, key: PublicKey) {
-        self.registered_keys.insert(key);
-    }
-}
-
-impl FeeHelper for EstimateFeesState {
-    type Error = WalletError;
-
-    fn account_exists(&self, key: &PublicKey) -> Result<bool, Self::Error> {
-        Ok(self.registered_keys.contains(key))
-    }
-}
-
-// State used to build a transaction
-// It contains the balances of the wallet and the registered keys
-pub struct TransactionBuilderState {
-    inner: EstimateFeesState,
-    mainnet: bool,
-    balances: HashMap<Hash, Balance>,
-    reference: Reference,
-    nonce: u64,
-}
-
-impl TransactionBuilderState {
-    pub fn new(mainnet: bool, reference: Reference, nonce: u64) -> Self {
-        Self {
-            inner: EstimateFeesState {
-                registered_keys: HashSet::new(),
-            },
-            mainnet,
-            balances: HashMap::new(),
-            reference,
-            nonce
-        }
-    }
-
-    pub fn set_balances(&mut self, balances: HashMap<Hash, Balance>) {
-        self.balances = balances;
-    }
-
-    pub fn add_balance(&mut self, asset: Hash, balance: Balance) {
-        self.balances.insert(asset, balance);
-    }
-
-    pub fn set_registered_keys(&mut self, registered_keys: HashSet<PublicKey>) {
-        self.inner.registered_keys = registered_keys;
-    }
-
-    pub fn add_registered_key(&mut self, key: PublicKey) {
-        self.inner.registered_keys.insert(key);
-    }
-}
-
-impl FeeHelper for TransactionBuilderState {
-    type Error = WalletError;
-
-    fn account_exists(&self, key: &PublicKey) -> Result<bool, Self::Error> {
-        self.inner.account_exists(key)
-    }
-}
-
-impl AccountState for TransactionBuilderState {
-    fn is_mainnet(&self) -> bool {
-        self.mainnet
-    }
-
-    fn get_reference(&self) -> Reference {
-        self.reference.clone()
-    }
-
-    fn get_account_balance(&self, asset: &Hash) -> Result<u64, Self::Error> {
-        self.balances.get(asset).map(|b| b.amount).ok_or_else(|| WalletError::BalanceNotFound(asset.clone()))
-    }
-
-    fn get_account_ciphertext(&self, asset: &Hash) -> Result<CiphertextCache, Self::Error> {
-        self.balances.get(asset).map(|b| b.ciphertext.clone()).ok_or_else(|| WalletError::BalanceNotFound(asset.clone()))
-    }
-
-    fn update_account_balance(&mut self, asset: &Hash, new_balance: u64, ciphertext: Ciphertext) -> Result<(), Self::Error> {
-        self.balances.insert(asset.clone(), Balance {
-            amount: new_balance,
-            ciphertext: CiphertextCache::Decompressed(ciphertext)
-        });
-        Ok(())
-    }
-
-    fn get_nonce(&self) -> Result<u64, Self::Error> {
-        Ok(self.nonce)
-    }
-
-    fn update_nonce(&mut self, new_nonce: u64) -> Result<(), Self::Error> {
-        self.nonce = new_nonce;
-        Ok(())
     }
 }
